memory file fix

+++ neo_body/memory.py
@@ -1,5 +1,6 @@
 from enum import Enum
 import sqlite3
+import os
 
 from agent import Agent
 
@@ -59,40 +60,41 @@ class Memory(Agent):
 
 
     def create_object_memory(self):
-        conn = sqlite3.connect('neo_test.db')
+        if not os.path.isfile('./neo_test.db'):
+            conn = sqlite3.connect('neo_test.db')
 
-        cursor = conn.cursor()
+            cursor = conn.cursor()
 
-        cursor.execute("""DROP TABLE IF EXISTS OBJECTS""")
-        cursor.execute("""CREATE TABLE IF NOT EXISTS OBJECTS
-                          (
-                           object_id INTEGER PRIMARY KEY ,
-                           OBJECT_NAME,
-                           OBJECT_COLOR,
-                           OBJECT_WEIGHT
-                           )""")
+            cursor.execute("""DROP TABLE IF EXISTS OBJECTS""")
+            cursor.execute("""CREATE TABLE IF NOT EXISTS OBJECTS
+                              (
+                               object_id INTEGER PRIMARY KEY ,
+                               OBJECT_NAME,
+                               OBJECT_COLOR,
+                               OBJECT_WEIGHT
+                               )""")
 
-        # create linking table between objects and adjectives
-        cursor.execute("""DROP TABLE IF EXISTS OBJECT_DESCRIPTION""")
-        cursor.execute("""CREATE TABLE IF NOT EXISTS OBJECT_DESCRIPTION
-                          (
-                           object_id INTEGER REFERENCES OBJECTS (object_id) ON DELETE CASCADE ,
-                           adjective_id INTEGER REFERENCES ADJECTIVES (adjective_id) ON DELETE CASCADE
-                          )""")
+            # create linking table between objects and adjectives
+            cursor.execute("""DROP TABLE IF EXISTS OBJECT_DESCRIPTION""")
+            cursor.execute("""CREATE TABLE IF NOT EXISTS OBJECT_DESCRIPTION
+                              (
+                               object_id INTEGER REFERENCES OBJECTS (object_id) ON DELETE CASCADE ,
+                               adjective_id INTEGER REFERENCES ADJECTIVES (adjective_id) ON DELETE CASCADE
+                              )""")
 
-        # creating adjective table
-        cursor.execute("""DROP TABLE IF EXISTS ADJECTIVES""")
-        cursor.execute("""CREATE TABLE IF NOT EXISTS ADJECTIVES
-                          (
-                           adjective_id INTEGER PRIMARY KEY ,
-                           adjective_name UNIQUE,
-                           category,
-                           less_than,
-                           greater_than
-                          )""")
+            # creating adjective table
+            cursor.execute("""DROP TABLE IF EXISTS ADJECTIVES""")
+            cursor.execute("""CREATE TABLE IF NOT EXISTS ADJECTIVES
+                              (
+                               adjective_id INTEGER PRIMARY KEY ,
+                               adjective_name UNIQUE,
+                               category,
+                               less_than,
+                               greater_than
+                              )""")
 
-        conn.commit()
-        conn.close()
+            conn.commit()
+            conn.close()
 
     def recall_objects(self):
         statement = self.ask("neo", "sql_statement")
